feat: Signal Files.app on FileProviderDomain change refactor: Split DriveInfoManager + FileProvider
